extmod/modussl_mbedtls: Fix polling during handshake and renegotiation
During the initial handshake or subsequent renegotiation, the protocol
might need to read in order to write (or conversely to write in order
to read). It might be blocked from doing so by the state of the
underlying socket (i.e. there is no data to read, or there is no space to write).

The library indicates this condition by returning `MBEDTLS_ERR_SSL_WANT_READ`
or `MBEDTLS_ERR_SSL_WANT_WRITE`. When that happens, we need to enforce
that the next poll operation only considers the direction that the library
indicated.