extmod/modussl_mbedtls: Wire in support for DTLS.
A new boolean argument `dtls` is added to `ussl.wrap_socket()`,
if true the library assumes that the underlying socket is a
datagram socket (i.e. UDP or similar).

Implement our own timer callbacks as the out of the box
implementation relies on `gettimeofday()`.

To fully support asyncio for DTLS socket, we will need to return
a readable or writable event in `poll(MP_STREAM_POLL, ...)` if
`_mbedtls_timing_get_delay(self) >= 1`. This is left for future
work so as not to interfere with #9871.